pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "win-gui-inspector"
version = "0.1.0"
description = "Windows GUI element inspector and mapper using pywinauto"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "obichan117" }]
keywords = ["windows", "gui", "inspector", "pywinauto", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Win32 (MS Windows)",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    "pywinauto>=0.6.9",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[project.scripts]
win-gui-inspector = "win_gui_inspector.cli:main"

[project.urls]
Repository = "https://github.com/obichan117/win-gui-inspector"

[tool.hatch.build.targets.wheel]
packages = ["src/win_gui_inspector"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "ARG", "SIM"]
ignore = ["E501", "B008"]
